requirements.txt
click==8.1.2
cycler==0.11.0
fonttools==4.32.0
joblib==1.1.0
kiwisolver==1.4.2
langdetect==1.0.9
matplotlib==3.5.1
mypy-extensions==0.4.3
nltk==3.7
numpy==1.22.3
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
pyparsing==3.0.8
python-dateutil==2.8.2
regex==2022.4.24
six==1.16.0
textblob==0.17.1
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.2.0
